fix(mason-lspconfig): We now ensure mason-lspconfig v2 attaches the mappings reliably. * No need for lsp.lua anymore (mason-lspconfig is 100% automatic). * The concept of lsp handlers has been removed from mason-lspconfig.
